Fix timestep alignment, add Gaussian splitting, LR decay, and Kaggle persistence
- Trainer: use dataset-corrected timesteps (split-index bug), exponential LR decay, periodic validation renders + PSNR, scaler-state checkpoints, keep-last-K pruning, empty_cache only at save boundaries

- Scheduler: real 3DGS-style Gaussian splitting (ellipsoid sampling, 1.6x scale reduction) alongside cloning, proportional budget thinning, single-pass row-mapped rebuilds keeping params/optimizer/is_dynamic aligned; grad threshold raised to 2e-4

- Dataset: optional scene normalization (center + 98th-pct radius into unit box), lstsq fallback for singular sparse camera rigs

- Losses: fp32 depth correlation; deformation MLP d_pos clamp

- Renderer/init: standard SH C0+0.5 color convention

- train.py: seeding, CUDA fail-fast, scene normalization wiring + scene_transform.json; evaluate.py applies the same transform

- Kaggle notebook: version-chaining checkpoint persistence, prebuilt gsplat wheel attempt, GPU/internet sanity cells, eval cell

- deps: torch>=2.3 + torchvision, pyproject aligned; README workflow corrected

+++ sparsechron/losses/depth.py
@@ -12,9 +12,11 @@ def depth_loss(pred_depth: torch.Tensor, gt_depth: torch.Tensor) -> torch.Tensor
     """
     assert isinstance(pred_depth, torch.Tensor), "pred_depth must be a torch.Tensor"
     assert isinstance(gt_depth, torch.Tensor), "gt_depth must be a torch.Tensor"
-    
-    pred_depth = pred_depth.squeeze(-1)
-    gt_depth = gt_depth.squeeze(-1)
+
+    # Correlation math in fp32: under autocast, fp16 variances over a full
+    # image can underflow and corrupt the Pearson ratio.
+    pred_depth = pred_depth.squeeze(-1).float()
+    gt_depth = gt_depth.squeeze(-1).float()
     
     valid_mask = (gt_depth > 0) & torch.isfinite(gt_depth)
     if not valid_mask.any():
